requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
requests
models